[ADD] 야생동물을 JSON파일을 통해서 생성할 수 있도록 변경
1. 이벤트 대신 Monobehaviour를 전달할 수 있도록 변경
2. 특수 Monobehaviour들 추가
- ERCharacterSkillMonobehaviour
- EREntityTimerMonobehaviour

+++ eternal_return/src/main/kotlin/org/eternalreturn/erentity/events/EREntityBurnEvent.kt
@@ -2,5 +2,5 @@ package org.eternalreturn.erentity.events
 
 import org.eternalreturn.erentity.EREntity
 
-class EREntityBurnEvent(val attacker : EREntity, val ticks : Int) : EREntityEvent {
+class EREntityBurnEvent(val attacker : EREntity, val ticks : Long) : EREntityEvent {
 }
